Pass connection wrapper around via pointer
Fixes #14

There was a finalization race in octod previously as vibe.d requires for all
connections to be released before connection pool is finalized. But because
connection wrapper was GC allocated, order of finalization was not
determenistic.

Simply adding destructor to `HTTPConnection` was not possible because it is
passed around by copy in API modules. To make it work, API method were changed
to accept/capture `HTTPConnection` by reference/pointer instead.